feature. Created post model
